+++ pom.xml
@@ -74,7 +74,7 @@
     <dependency>
       <groupId>com.fasterxml.jackson.core</groupId>
       <artifactId>jackson-databind</artifactId>
-      <version>2.13.3</version>
+      <version>2.15.3</version>
     </dependency>
 
     <dependency>
