+++ src/main/java/com/airbng/dto/reservation/ReservationSearchResponse.java
@@ -1,4 +1,4 @@
-package com.airbng.dto;
+package com.airbng.dto.reservation;
 
 import lombok.*;
